requirements.txt
anyio==2.1.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
future==0.18.2
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.0
json5==0.9.5
jsondiff==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
linecache2==1.0.0
lockfile==0.12.2
luigi==3.0.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numpy==1.19.5
packaging==20.9
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
Pebble==4.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip==20.3.3
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
pyzmq==20.0.0
qgrid==1.3.1
regex==2020.11.13
requests==2.25.1
Send2Trash==1.5.0
setuptools==51.1.2.post20210110
six==1.15.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
timeout-decorator==0.5.0
tornado==6.1
tqdm==4.56.0
traceback2==1.4.0
traitlets==5.0.5
treelib==1.6.1
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
xlrd==1.2.0
zipp==3.4.0
